added character movement rendering and fps

+++ src/application/controller/CharacterController.java
@@ -2,6 +2,6 @@ package application.controller;
 
 public abstract class CharacterController extends GameObjectController{
 	
-	public abstract void move(int x, int y);
+	public abstract void move(double x, double y);
 
 }
